Challenge5/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-cloud-bigquery==3.14.1
google-cloud-logging==3.9.0
google-genai>=1.0.0
google-cloud-aiplatform==1.38.1
pandas>=2.0.0
pytest>=7.4.0